# speed up missing data imputation